* Refactoring of multiple file deletion * Bugfix in fileindex

+++ inc/controller/ajax/files/delete.php
@@ -25,9 +25,18 @@ class delete extends \fpcm\controller\abstracts\ajaxController {
 
     /**
      *
-     * @var string
+     * @var bool
      */
-    private $newFileName = '';
+    private $multiple = false;
+
+    /**
+     *
+     * @var array
+     */
+    private $deleted = [
+        'ok' => [],
+        'failed' => [],
+    ];
 
     /**
      * 
@@ -48,6 +57,7 @@ class delete extends \fpcm\controller\abstracts\ajaxController {
             \fpcm\classes\http::FILTER_BASE64DECODE
         ]);
 
+        $this->multiple = (bool) $this->getRequestVar('multiple');
         return true;
     }
 
@@ -56,18 +66,65 @@ class delete extends \fpcm\controller\abstracts\ajaxController {
      */
     public function process()
     {
-        $replace = ['{{filename1}}' => basename($this->fileName), '{{filename2}}' => basename($this->newFileName)];
+        call_user_func([$this, 'delete'.($this->multiple ? 'Multiple' : 'Single')]);
+        $this->getSimpleResponse();
+    }
+    
+    private function deleteSingle()
+    {
+        if (!is_string($this->fileName) || !trim($this->fileName)) {
+            $this->returnData['code'] = 0;
+            $this->returnData['message'] = $this->language->translate('DELETE_FAILED_FILES', '');
+            return false;
+        }
         
+        $replace = ['{{filenames}}' => basename($this->fileName)];
+
         $img = new \fpcm\model\files\image($this->fileName, false);
         if ($img->isValidDataFolder('', \fpcm\classes\dirs::DATA_UPLOADS) && $img->delete()) {
             $this->returnData['code'] = 1;
             $this->returnData['message'] = $this->language->translate('DELETE_SUCCESS_FILES', $replace);
-            $this->getSimpleResponse();
+            return true;
         }
 
         $this->returnData['code'] = 0;
         $this->returnData['message'] = $this->language->translate('DELETE_FAILED_FILES', $replace);
-        $this->getSimpleResponse();
+        return false;
+    }
+    
+    private function deleteMultiple()
+    {
+        if (!is_array($this->fileName) || !count($this->fileName)) {
+            $this->returnData['code'][1] = 0;
+            $this->returnData['message'][1] = $this->language->translate('DELETE_FAILED_FILES', '');
+            return false;
+        }
+
+        array_walk($this->fileName, function ($fileName)
+        {
+            if ((new \fpcm\model\files\image($fileName, false))->delete()) {
+                $this->deleted['ok'][] = $fileName;
+                return true;
+            }
+
+            $this->deleted['failed'][] = $fileName;
+            return false;
+        });
+
+        if (count($this->deleted['ok'])) {
+            $this->returnData['code'][1] = 1;
+            $this->returnData['message'][1] = $this->language->translate('DELETE_SUCCESS_FILES', [
+                '{{filenames}}' => implode(', ', $this->deleted['ok'])
+            ]);
+        }
+
+        if (count($this->deleted['failed'])) {
+            $this->returnData['code'][2] = 0;
+            $this->returnData['message'][2] = $this->language->translate('DELETE_FAILED_FILES', [
+                '{{filenames}}' => implode(', ', $this->deleted['failed'])
+            ]);            
+        }
+
     }
 
 }
